+++ src/main/java/kr/codesquad/Lotto/Prize.java
@@ -1,4 +1,4 @@
-package kr.codesquad;
+package kr.codesquad.Lotto;
 
 public enum Prize {
     FIRST(2000000000), SECOND(3000000), THIRD(1500000), FORTH(50000), FIFTH(5000), NOTHING(0);
