--- src/main/java/kr/codesquad/Main.java
@@ -1,13 +1,0 @@
-package kr.codesquad;
-
-import java.util.List;
-import java.util.ArrayList;
-
-public class Main {
-    public static void main(String[] args) {
-        LottoGame lottoGame = new LottoGame();
-        lottoGame.start();
-    }
-
-
-}
